+++ websockets/client.py
@@ -6,7 +6,7 @@ The :mod:`websockets.client` module defines a simple WebSocket client API.
 import asyncio
 import collections.abc
 
-from .exceptions import InvalidHandshake, InvalidMessage
+from .exceptions import InvalidHandshake, InvalidMessage, InvalidStatus
 from .handshake import build_request, check_response
 from .http import USER_AGENT, build_headers, read_response
 from .protocol import CONNECTING, OPEN, WebSocketCommonProtocol
@@ -56,14 +56,15 @@ class WebSocketClientProtocol(WebSocketCommonProtocol):
 
         """
         try:
-            status_code, headers = yield from read_response(self.reader)
+            status_code, headers, reason = (
+                yield from read_response(self.reader))
         except ValueError as exc:
             raise InvalidMessage("Malformed HTTP message") from exc
 
         self.response_headers = build_headers(headers)
         self.raw_response_headers = headers
 
-        return status_code, self.response_headers
+        return status_code, self.response_headers, reason
 
     def process_subprotocol(self, get_header, subprotocols=None):
         """
@@ -114,11 +115,11 @@ class WebSocketClientProtocol(WebSocketCommonProtocol):
 
         yield from self.write_http_request(wsuri.resource_name, headers)
 
-        status_code, headers = yield from self.read_http_response()
+        status_code, headers, reason = yield from self.read_http_response()
         get_header = lambda k: headers.get(k, '')
 
         if status_code != 101:
-            raise InvalidHandshake("Bad status code: {}".format(status_code))
+            raise InvalidStatus(status_code, reason=reason)
 
         check_response(get_header, key)