+++ websockets/test_http.py
@@ -44,9 +44,11 @@ class HTTPAsyncTests(unittest.TestCase):
             b'Sec-WebSocket-Protocol: chat\r\n'
             b'\r\n'
         )
-        status, hdrs = self.loop.run_until_complete(read_response(self.stream))
+        status, headers, reason = self.loop.run_until_complete(
+            read_response(self.stream))
         self.assertEqual(status, 101)
-        self.assertEqual(dict(hdrs)['Upgrade'], 'websocket')
+        self.assertEqual(dict(headers)['Upgrade'], 'websocket')
+        self.assertEqual(reason, 'Switching Protocols')
 
     def test_request_method(self):
         self.stream.feed_data(b'OPTIONS * HTTP/1.1\r\n\r\n')